% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_github_repositories.R
\name{search_github_repositories}
\alias{search_github_repositories}
\title{search_github_repositories}
\usage{
search_github_repositories(query, sort = "stars", order = "desc")
}
\arguments{
\item{query}{Search query}

\item{sort}{Sort}

\item{order}{Order result}
}
\value{
Parsed response
}
\description{
search_github_repositories
}
\examples{
search_github_repositories(query = "sta-518")
}
